Move some methods from the fast path back to the default
